refactor: move subscription into own module
