+++ src/event.rs
@@ -3,9 +3,11 @@ pub mod event;
 pub mod event_repeater;
 pub mod observable;
 pub mod subscriber;
+pub mod subscription;
 
 pub use arc_observable::ArcObservable;
 pub use event::Event;
 pub use event_repeater::EventRepeater;
 pub use observable::{Observable, ObservableResult};
 pub use subscriber::{Callback, DispatchError, Subscriber};
+pub use subscription::{ReceiverSubscription, Subscription};
